FIX: validation and error helpers have a Nothing return type
These helpers used to have a Unit return type, making them cumbersome to
use in places where the return type was of another type. Since these
return a validation or error failure variant of the ADT, the return type
does not matter, therefore it should become the bottom type Nothing.